Fix for #876
The notifications really can't work under Oreo, because you make the channel with some strange id, while providing the id as the name in the next argument. Maybe not the best fix, but I find this string to be as cool of an id as it is of a name. Definitely better than having no notifications at all.